Make the discoveries read callable without an HTTP trip
The whole read lived inside the route handler, so the only way to reach it was
a request — a server component wanting the same facts had to fetch its own
process over the network and miss the single-flight memo entirely. On a chain
already refusing this fleet at ~442 rate-limit hits in five minutes, a second
upstream read for facts we are holding in memory is the expensive mistake, and
the token page needs exactly those facts.

lib/read-discoveries now owns the read and the memo; the route is a thin GET
over it. Two components were importing their types from the route file, which
is the smell that motivated this; they import from the reader now.

honesty.test.ts scans both halves rather than picking one. Its assertions are
about the discoveries read as a whole, and filing each under a half would mean
deciding per assertion which side may satisfy it — the first one filed wrong
would pass by accident forever.
