+++ web/src/app/api/discoveries/route.ts
@@ -1,377 +1,18 @@
 import { NextResponse } from "next/server";
-import { createPublicClient, http } from "viem";
-import {
-  fetchGeckoPoolsResult,
-  screenPools,
-  type GeckoPool,
-} from "../../../../../worker/src/venues/geckoterminal";
-import { recentPonsLaunches } from "../../../../../worker/src/venues/pons";
-import {
-  readCurveActivity,
-  isActive,
-  MAX_ACTIVITY_BLOCKS,
-} from "../../../../../worker/src/venues/pons-activity";
-import { readTokenMeta } from "../../../../../worker/src/venues/pons-meta";
-import {
-  readCardFacts,
-  readBlockClock,
-  ageSecOf,
-} from "../../../../../worker/src/venues/pons-card";
-import { resolveConfig } from "../../../../../worker/src/settings";
-import { resolveLlm } from "../../../../../worker/src/llm";
-import { createMemecoinScout } from "../../../../../worker/src/strategist/memecoin-scout";
-import { researchCoins, scoutFieldsFor } from "../../../../../worker/src/strategist/coin-research";
+import { sharedRead } from "@/lib/read-discoveries";
 
 /**
- * What is trading on this chain, for the dashboard.
+ * What is trading on this chain, over HTTP.
  *
- * READS THE INDEX, NOT THE LEDGER, and that is the whole design decision.
- * Every other panel here goes through `withReadDb` against the worker's
- * database — which works self-hosted and renders EMPTY on the hosted deploy,
- * because the orchestrator strips DATABASE_URL from each worker child, so the
- * child writes sqlite in its own container while this service reads a Postgres
- * nothing ever created the schema in. A discoveries panel built the same way
- * would be blank on app.merrymen.dev for exactly that reason.
- *
- * Fetching server-side instead means the panel shows the same thing to
- * everyone, hosted or not, with no database and no tenant scoping —
- * `discovered_pools` has no tenant column anyway. It is the same shape
- * /api/market already uses.
- *
- * WHAT THIS IS. A third party's claim about a market, used to decide what is
- * worth LOOKING at. It is not a recommendation, nothing here has been checked
- * against the chain, and the agent cannot act on any of it without the owner
- * adding the token in /settings and re-signing the grant.
+ * The read itself lives in lib/read-discoveries so a server component can call
+ * it directly and share the single-flight memo, rather than fetching this
+ * route and paying a second trip upstream. Everything of substance — what is
+ * fetched, what is refused, what "degraded" means — is documented there.
  */
-
 // Dynamic, with an in-process single-flight memo below instead of ISR: this
 // route must be able to REFUSE to cache a degraded render, which a fixed
 // revalidate cannot do.
 export const dynamic = "force-dynamic";
-
-/** The same floor the worker's own screen uses, so the two agree on "worth showing". */
-const LIMITS = { minReserveUsd: 25_000, minVolume24hUsd: 50_000, minBuyers24h: 100 };
-
-/** GeckoTerminal's venue slugs for the two halves of the Pons launchpad. */
-const GRADUATED = "pons-v2-dex";
-const ON_CURVE = "pons-v2";
-
-export interface DiscoveryRow {
-  token: string;
-  name: string;
-  venue: string;
-  priceUsd: number | null;
-  /**
-   * Depth as the INDEX reports it.
-   *
-   * For a coin still on its bonding curve this is mostly the VIRTUAL SEED — a
-   * fresh curve reports about $4,100 of "reserve" while holding none of it —
-   * so the UI must never present it as money you could sell into. That is why
-   * `onCurve` travels with it.
-   */
-  reserveUsd: number | null;
-  fdvUsd: number | null;
-  volume24hUsd: number | null;
-  change24hPct: number | null;
-  buyers24h: number | null;
-  ageDays: number | null;
-  /** Graduated off a Pons bonding curve into a real pool. */
-  graduated: boolean;
-  /** Still on its bonding curve — treat `reserveUsd` with suspicion. */
-  onCurve: boolean;
-  /**
-   * The scout's own line on this coin, when it had one.
-   *
-   * Declared here because the builder has always ATTACHED it while the type
-   * never admitted it — so every consumer had to re-declare this interface
-   * locally, which is exactly how the console ended up with a private copy of
-   * it. Null is a real answer: the scout looked and passed. `verdictsWhy` on
-   * the payload says whether it could look at all.
-   */
-  verdict?: { conviction: number; reason: string } | null;
-}
-
-function toRow(p: GeckoPool, nowSec: number): DiscoveryRow {
-  return {
-    token: p.tokenAddress,
-    // The index's own label, shown as a label and never used as identity: the
-    // worker reads a symbol from the contract precisely because this string is
-    // attacker-chosen and could impersonate a real ticker.
-    name: p.name,
-    venue: p.dex,
-    priceUsd: p.priceUsd,
-    reserveUsd: p.reserveUsd,
-    fdvUsd: p.fdvUsd,
-    volume24hUsd: p.volume24hUsd,
-    change24hPct: p.change24hPct,
-    buyers24h: p.buyers24h,
-    ageDays: p.createdAt === null ? null : Math.max(0, (nowSec - p.createdAt) / 86_400),
-    graduated: p.dex === GRADUATED,
-    onCurve: p.dex === ON_CURVE,
-  };
-}
-
-/** A launch from the last few minutes that people are actually trading. */
-export interface FreshRow {
-  token: string;
-  curve: string;
-  trades: number;
-  /** Distinct trading ADDRESSES, from the trade event's own indexed field. */
-  traders: number;
-  /** The launcher's own words. Sanitised, and a claim rather than a fact. */
-  description: string;
-  twitter: string;
-  telegram: string;
-  website: string;
-  /** Published nothing at all — the shape an abandoned template has. */
-  bare: boolean;
-  /** The ERC-20's own ticker, from the chain. Empty when unreadable. */
-  symbol: string;
-  /** The ERC-20's own name, from the chain. Empty when unreadable. */
-  name: string;
-  /**
-   * The launcher's logo URI, usually `ipfs://…`. Never given to a browser
-   * directly — it goes through /api/coin-image, which is both what makes it
-   * load at all (every gateway 403s a browser User-Agent) and what keeps an
-   * attacker-chosen URL out of the reader's browser.
-   */
-  logo: string;
-  /** Seconds since it launched, measured from a real block clock. Null when unknown. */
-  ageSec: number | null;
-  /** Basis points of its own graduation threshold, net of the virtual seed. */
-  progressBps: number | null;
-}
-
-/**
- * The other end of the launchpad: what launched in the last quarter hour and
- * has a tape.
- *
- * Pons runs at roughly 940 launches an hour, so this is a FUNNEL rather than a
- * list. The gate is trading — 25 trades and 3 distinct addresses — which keeps
- * about an eighth of launches and holds 96% of the ones that go on to graduate.
- * A dev buy, having socials, and the creator's history were all measured and
- * are worth nothing as filters.
- *
- * Three RPC calls for the whole thing, whatever the launch rate: the launches,
- * one chain-wide sweep of every curve trade, and one Multicall3 batch for the
- * survivors' metadata.
- */
-/**
- * Which of the three enrichment reads came back.
- *
- * These fail as a WAVE, not one coin at a time — the RPC refuses the burst and
- * all three return nothing together — so the answer belongs to the page, not to
- * a card. Observed in production on 2026-08-30: 28 rows with real trade counts
- * and every enriched field blank, served for about two and a half minutes.
- */
-export interface ChainStatus {
-  /**
-   * The launch scan and the trade sweep — the reads that produce the ROWS.
-   *
-   * Separate from the three below because it fails differently and worse: when
-   * this goes, there are no rows at all, and an empty list renders as "nothing
-   * launched in the last few minutes has anyone trading it" — a confident claim
-   * about a launchpad running at 940 launches an hour. Same mistake as `bare`,
-   * one level up.
-   */
-  launchpad: boolean;
-  /** The launcher's description, logo and socials (pons-meta). */
-  meta: boolean;
-  /** Symbol, name and curve progress (pons-card). */
-  facts: boolean;
-  /** The block clock that turns a block number into an age. */
-  clock: boolean;
-}
-
-/** Space out a burst: the reads are cheap, the RPC's tolerance for concurrency is not. */
-const sleep = (ms: number) => new Promise((r) => setTimeout(r, ms));
-
-/**
- * Rank the screened set with the same scout the worker uses.
- *
- * Optionally researches first, when a browser is configured on THIS service —
- * the signals then reach the model exactly as they do in the worker. Absent a
- * browser it ranks on numbers alone, which is the worker's behaviour too.
- *
- * Never throws: a page that cannot form an opinion still shows the market. And
- * with no model configured it returns no picks at all rather than a default
- * opinion — "nothing has been vetted" is the honest answer there, not
- * "everything looks fine".
- */
-async function rankForDisplay(
-  kept: readonly GeckoPool[],
-  nowSec: number,
-): Promise<{
-  picks: { pool: GeckoPool; conviction: number; reason: string }[];
-  /**
-   * Why there are no verdicts, when there are none.
-   *
-   * "no-model", "model-failed" and "chose nothing" are three different facts and
-   * only the last is an opinion. The scout fails CLOSED on a provider error —
-   * correctly, since its whole job is to cut a list down — but that makes an
-   * outage indistinguishable from a considered pass, and a page showing no
-   * verdicts would read as "the agent looked and liked nothing" either way.
-   * That is the same silent-failure shape as the coin cards and the market
-   * index, so it gets the same treatment.
-   */
-  why?: "no-model" | "model-failed";
-}> {
-  if (!kept.length) return { picks: [] };
-  let creds;
-  let cfg;
-  try {
-    cfg = resolveConfig();
-    creds = resolveLlm(cfg);
-  } catch {
-    return { picks: [], why: "no-model" };
-  }
-  if (!creds) return { picks: [], why: "no-model" };
-  try {
-
-    let research: ReadonlyMap<string, ReturnType<typeof scoutFieldsFor>> | undefined;
-    if (cfg.browserUrl && cfg.browserToken) {
-      const client = createPublicClient({ transport: http("https://rpc.mainnet.chain.robinhood.com") });
-      const found = await researchCoins(kept, {
-        client: client as never,
-        browser: { baseUrl: cfg.browserUrl, token: cfg.browserToken },
-      });
-      research = new Map([...found].map(([k, r]) => [k, scoutFieldsFor(r)]));
-    }
-    return await createMemecoinScout(creds).rank(kept, nowSec, research);
-  } catch {
-    return { picks: [] };
-  }
-}
-
-async function readFresh(): Promise<{ rows: FreshRow[]; chain: ChainStatus }> {
-  // `true` means "read it", so a total failure below reports three honest
-  // falses rather than three optimistic trues.
-  const chain: ChainStatus = { launchpad: false, meta: false, facts: false, clock: false };
-  try {
-    const client = createPublicClient({ transport: http("https://rpc.mainnet.chain.robinhood.com") });
-    const W = MAX_ACTIVITY_BLOCKS;
-    const [scan, activity] = await Promise.all([
-      recentPonsLaunches(client as never, W),
-      readCurveActivity(client as never, W),
-    ]);
-    // A null activity map means the node refused, which is a different fact
-    // from a quiet launchpad — showing nothing is right, inventing an empty
-    // tape for every launch is not.
-    if (scan.failed || !activity) return { rows: [], chain };
-    chain.launchpad = true;
-    const live = scan.launches.filter((l) => isActive(activity.get(l.curve.toLowerCase())));
-
-    // SEQUENTIAL, NOT Promise.all — and this is the one behavioural change here.
-    // These three used to fire as a burst immediately after two heavy log
-    // sweeps, and the whole burst is what the node refuses: all three come back
-    // empty together while the sweeps that preceded them succeeded. They cost
-    // ~700ms in total, so spacing them is nearly free, and the alternative
-    // (more retries) multiplies the burst that draws the refusal.
-    const meta = await readTokenMeta(client as never, live.map((l) => l.token));
-    chain.meta = meta.size > 0 || live.length === 0;
-    await sleep(120);
-    const facts = await readCardFacts(client as never, live);
-    chain.facts = facts.size > 0 || live.length === 0;
-    await sleep(120);
-    const clock = await readBlockClock(client as never);
-    chain.clock = clock !== null;
-
-    return {
-      chain,
-      rows: live
-        .map((l) => {
-          const a = activity.get(l.curve.toLowerCase())!;
-          const m = meta.get(l.token.toLowerCase());
-          const f = facts.get(l.token.toLowerCase());
-          return {
-            token: l.token,
-            curve: l.curve,
-            trades: a.buys + a.sells,
-            traders: a.traders,
-            description: m?.description ?? "",
-            twitter: m?.twitter ?? "",
-            telegram: m?.telegram ?? "",
-            website: m?.website ?? "",
-            // `bare` means "this launcher published NOTHING", which is a claim
-            // about the coin — so it may only be made when the read succeeded.
-            // It used to be `m ? m.bare : true`, which turned every unread coin
-            // into an accusation: the card said "Published nothing about
-            // itself" and "no socials" about coins that published plenty.
-            // readTokenMeta returns a MAP precisely so a caller can tell "read
-            // and empty" from "not read", and this threw that away.
-            bare: m ? m.bare : false,
-            symbol: f?.symbol ?? "",
-            name: f?.name ?? "",
-            logo: m?.logo ?? "",
-            ageSec: ageSecOf(clock, l.blockNumber),
-            progressBps: f?.progressBps ?? null,
-          };
-        })
-        // By distinct addresses, not trade count: 291 trades from 25 addresses is
-        // a different thing from 223 trades from 176, and only one of them looks
-        // like people.
-        .sort((x, y) => y.traders - x.traders),
-    };
-  } catch {
-    return { rows: [], chain };
-  }
-}
-
-export interface Payload {
-  fetchedAt: number;
-  scanned: number;
-  indexUnreachable: boolean;
-  rows: DiscoveryRow[];
-  graduated: number;
-  fresh: FreshRow[];
-  chain: ChainStatus;
-  /**
-   * Why no coin carries a verdict, when none does.
-   *
-   * null is a real opinion — the scout looked and picked nothing, which the
-   * prompt says is often the right answer. A value means it could not look, and
-   * the page must not render that as a considered pass.
-   */
-  verdictsWhy: "no-model" | "model-failed" | null;
-  degraded: boolean;
-}
-
-/**
- * One in-flight read at a time, and one result shared by every viewer.
- *
- * REPLACES `export const revalidate = 120`, which cached whatever it got —
- * including a degraded render, which it then served for minutes. This is the
- * same sharing with a say in what gets kept: a whole read is reused for two
- * minutes, a degraded one for ten seconds.
- *
- * The single-flight part is not optional once the route is dynamic. The console
- * polls this every 120s PER OPEN TAB; without it, every tab that misses fires
- * two heavy log sweeps plus three enrichment reads into a keyless RPC — which is
- * precisely the burst that makes the enrichment fail in the first place.
- *
- * A per-process memo is enough while `web` runs one replica (railway.json sets
- * no replica count). Scale it and each replica keeps its own — still correct,
- * just N times the upstream traffic.
- */
-let inFlight: Promise<Payload> | null = null;
-let last: { at: number; payload: Payload } | null = null;
-const WHOLE_MS = 120_000;
-const DEGRADED_MS = 10_000;
-
-function sharedRead(): Promise<Payload> {
-  const ttl = last?.payload.degraded ? DEGRADED_MS : WHOLE_MS;
-  if (last && Date.now() - last.at < ttl) return Promise.resolve(last.payload);
-  if (inFlight) return inFlight;
-  inFlight = build()
-    .then((p) => {
-      last = { at: Date.now(), payload: p };
-      return p;
-    })
-    .finally(() => {
-      inFlight = null;
-    });
-  return inFlight;
-}
 
 export async function GET() {
   const payload = await sharedRead();
@@ -383,93 +24,4 @@ export async function GET() {
         : "public, s-maxage=120, stale-while-revalidate=240",
     },
   });
-}
-
-async function build(): Promise<Payload> {
-  const nowSec = Math.floor(Date.now() / 1000);
-  const { rows: fresh, chain } = await readFresh();
-  const byToken = new Map<string, GeckoPool>();
-  // Every feed refused is a different fact from every feed being empty, and
-  // only one of them means the market is quiet. This API is keyless and
-  // rate-limited, so the refusal is routine — and a page that renders it as
-  // "nothing clearing the floor" states something false while looking normal.
-  let asked = 0;
-  let reached = 0;
-  for (const feed of ["trending_pools", "new_pools", "pools"] as const) {
-    const r = await fetchGeckoPoolsResult(feed);
-    asked++;
-    if (!r.failed) reached++;
-    for (const p of r.pools) {
-      // Deduped by TOKEN and kept at its deepest venue — the same coin appears
-      // in several feeds and often on several venues, and a reader cares about
-      // the coin.
-      const prev = byToken.get(p.tokenAddress);
-      if (!prev || (p.reserveUsd ?? 0) > (prev.reserveUsd ?? 0)) byToken.set(p.tokenAddress, p);
-    }
-  }
-  const all = [...byToken.values()];
-  const { kept } = screenPools(all, LIMITS);
-
-  // ── THE AGENT'S OWN VERDICT, FORMED HERE ────────────────────────────────
-  //
-  // Server-side for the same reason the screen above is: a worker child has no
-  // DATABASE_URL (the orchestrator strips it), so it writes its ledger to sqlite
-  // in its own container and NOTHING it decides can be read by this service.
-  // A verdict panel fed from `decisions` would be permanently empty on
-  // app.merrymen.dev — the exact failure this route's header was written about.
-  //
-  // So the same model that ranks candidates in the worker ranks them again
-  // here, on the same screened set, and the answer travels with the row. It is
-  // one LLM call per render, shared by every viewer through the memo above.
-  //
-  // This is a READING, not a permission. Nothing here can authorise a trade:
-  // the wall only admits assets sealed into a signature, and the scout is
-  // structurally incapable of naming a token it was not offered.
-  const scoutRes = await rankForDisplay(kept, nowSec);
-  const verdictsWhy = scoutRes.why ?? null;
-  const verdictByToken = new Map<string, { conviction: number; reason: string }>();
-  for (const pick of scoutRes.picks) {
-    verdictByToken.set(pick.pool.tokenAddress.toLowerCase(), {
-      conviction: pick.conviction,
-      reason: pick.reason,
-    });
-  }
-
-  const rows = kept.map((p) => ({
-    ...toRow(p, nowSec),
-    verdict: verdictByToken.get(p.tokenAddress.toLowerCase()) ?? null,
-  }));
-  // Graduated first, then by 24h move: a coin that just made it off the
-  // launchpad is the thing this page exists to surface.
-  rows.sort((a, b) => Number(b.graduated) - Number(a.graduated) || (b.change24hPct ?? 0) - (a.change24hPct ?? 0));
-
-  // A render worth keeping is one where every source answered. Anything less
-  // gets a short life, so the next viewer re-asks instead of inheriting it.
-  //
-  // CACHING IS WHAT TURNED A BLINK INTO AN OUTAGE. The enrichment reads fail as
-  // an occasional wave and recover in seconds, but a degraded render used to be
-  // cached like any other: measured in production, one bad render was served
-  // `x-nextjs-cache: HIT` for six consecutive polls — about two and a half
-  // minutes of every card claiming its coin had published nothing. Only the
-  // cache made it last that long. A render worth keeping is one where every
-  // source answered; anything less gets a short life, so the next viewer
-  // re-asks instead of inheriting it.
-  const degraded =
-    !chain.launchpad || !chain.meta || !chain.facts || !chain.clock || (reached === 0 && asked > 0);
-
-  return {
-    fetchedAt: nowSec,
-    scanned: all.length,
-    // False only when NOT ONE feed answered. A partial read is still a read.
-    indexUnreachable: reached === 0 && asked > 0,
-    rows,
-    graduated: rows.filter((r) => r.graduated).length,
-    fresh,
-    chain,
-    // Null means the scout looked and picked nothing, which is a real and often
-    // correct answer. A value means it could not look at all — and that must
-    // not read on the page as a considered pass.
-    verdictsWhy,
-    degraded,
-  };
 }
